# docs-site/openapi.yaml
openapi: 3.0.3
info:
  title: SwipeSavvy Backend API
  description: |
    Multi-service backend API for SwipeSavvy Mobile Wallet — a fintech platform providing
    card issuance (FIS), payments (Authorize.Net), KYC, rewards, real-time chat, and admin management.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained via `POST /api/v1/auth/login` + `POST /api/v1/auth/verify-login-otp`.

    | Token | Lifetime | Purpose |
    |-------|----------|---------|
    | Access Token | 24 hours | Authorize API requests |
    | Refresh Token | 30 days | Obtain new access tokens |

    ## Environments
    Use the server selector above to switch between Production, Staging, and Development.
  version: 1.0.0
  contact:
    name: SwipeSavvy Engineering
    email: engineering@swipesavvy.com
  license:
    name: Proprietary
servers:
  - url: http://54.224.8.14:8000
    description: Production (AWS EC2)
  - url: https://api.swipesavvy.com
    description: Production (Domain)
  - url: http://localhost:8000
    description: Development (Local)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from login flow
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT token from admin login
    WebhookSignature:
      type: apiKey
      in: header
      name: X-FIS-Signature
      description: FIS webhook signature verification
  schemas:
...
    FeatureFlag:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        rollout_percentage:
          type: integer
          minimum: 0
          maximum: 100
        environment:
          type: string
        created_at:
          type: string
          format: date-time
...
paths:
...
  /api/feature-flags:
    post:
      tags: [Feature Flags]
      operationId: createFeatureFlag
      summary: Create a feature flag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                enabled:
                  type: boolean
                  default: false
                rollout_percentage:
                  type: integer
                  default: 0
                environment:
                  type: string
                  default: development
      responses:
        '200':
          description: Flag created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/FeatureFlag'
    get:
      tags: [Feature Flags]
      operationId: listFeatureFlags
      summary: List all feature flags
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 50
        - in: query
          name: enabled_only
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Flag list
  /api/feature-flags/{flag_id}:
    get:
      tags: [Feature Flags]
      operationId: getFeatureFlag
      summary: Get feature flag by ID
      parameters:
        - in: path
          name: flag_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Flag details
    put:
      tags: [Feature Flags]
      operationId: updateFeatureFlag
      summary: Update a feature flag
      parameters:
        - in: path
          name: flag_id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                enabled:
                  type: boolean
                rollout_percentage:
                  type: integer
      responses:
        '200':
          description: Flag updated
    delete:
      tags: [Feature Flags]
      operationId: deleteFeatureFlag
      summary: Delete a feature flag
      parameters:
        - in: path
          name: flag_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Flag deleted
  /api/feature-flags/name/{flag_name}:
    get:
      tags: [Feature Flags]
      operationId: getFeatureFlagByName
      summary: Get feature flag by name
      parameters:
        - in: path
          name: flag_name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Flag details
  /api/feature-flags/{flag_id}/toggle:
    patch:
      tags: [Feature Flags]
      operationId: toggleFeatureFlag
      summary: Toggle feature flag on/off
      parameters:
        - in: path
          name: flag_id
          required: true
          schema:
            type: string
        - in: query
          name: enabled
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Flag toggled
  /api/feature-flags/mobile/active:
    get:
      tags: [Feature Flags]
      operationId: getMobileActiveFlags
      summary: Get active flags for mobile app
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
      responses:
        '200':
          description: Active mobile flags

  # ═══════════════════════════════════════════
  # ADMIN AUTH (7 endpoints)
  # ═══════════════════════════════════════════
...
  /api/v1/admin/feature-flags:
    post:
      tags: [Admin Feature Flags]
      operationId: adminCreateFlag
      summary: Create admin feature flag
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                enabled:
                  type: boolean
                rollout_percentage:
                  type: integer
                category:
                  type: string
      responses:
        '200':
          description: Flag created
    get:
      tags: [Admin Feature Flags]
      operationId: adminListFlags
      summary: List admin feature flags
      security:
        - AdminAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: per_page
          schema:
            type: integer
        - in: query
          name: enabled
          schema:
            type: boolean
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: Flag list
  /api/v1/admin/feature-flags/{flag_id}:
    get:
      tags: [Admin Feature Flags]
      operationId: adminGetFlag
      summary: Get flag details
      security:
        - AdminAuth: []
      parameters:
        - in: path
          name: flag_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Flag details
    put:
      tags: [Admin Feature Flags]
      operationId: adminUpdateFlag
      summary: Update flag
      security:
        - AdminAuth: []
      parameters:
        - in: path
          name: flag_id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                enabled:
                  type: boolean
                rollout_percentage:
                  type: integer
      responses:
        '200':
          description: Flag updated
    delete:
      tags: [Admin Feature Flags]
      operationId: adminDeleteFlag
      summary: Delete flag
      security:
        - AdminAuth: []
      parameters:
        - in: path
          name: flag_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Flag deleted
  /api/v1/admin/feature-flags/{flag_id}/toggle:
    put:
      tags: [Admin Feature Flags]
      operationId: adminToggleFlag
      summary: Toggle flag on/off
      security:
        - AdminAuth: []
      parameters:
        - in: path
          name: flag_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        '200':
          description: Flag toggled
...
  /api/v1/location/route:
    post:
      tags: [Location]
      operationId: calculateRoute
      summary: Calculate route between points
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [origin, destination]
              properties:
                origin:
                  type: object
                  properties:
                    latitude:
                      type: number
                    longitude:
                      type: number
                destination:
                  type: object
                  properties:
                    latitude:
                      type: number
                    longitude:
                      type: number
                mode:
                  type: string
                  enum: [driving, walking, cycling]
      responses:
        '200':
          description: Route details
...